Safari-609.1.11.1/LayoutTests/js/dom/array-tostring-ignore-separator-expected.txt

This test checks for regressions against 6261:Do not use a separator argument when doing toString/toLocalString.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a.toString('!') is '5,3'
PASS successfullyParsed is true

TEST COMPLETE
